+++ app/src/main/java/com/zetung/gifsgiver/api/GifApi.kt
@@ -1,0 +1,9 @@
+package com.zetung.gifsgiver.api
+
+import com.zetung.gifsgiver.model.AllGifs
+import retrofit2.http.GET
+
+interface GifApi {
+    @GET("gifs/trending?api_key=XR5yDta2h0usLLHQDcHW6LUI8nI6WtMy&limit=25&offset=0&rating=g&bundle=messaging_non_clips")
+    fun getGifs(): retrofit2.Call<AllGifs>
+}
