--- app/src/main/java/com/zetung/gifsgiver/model/AllGifs.kt
@@ -1,5 +1,0 @@
-package com.zetung.gifsgiver.model
-
-data class AllGifs(
-    val gifs: List<Gif>
-)
