--- app/src/main/java/com/zetung/gifsgiver/model/Gif.kt
@@ -1,5 +1,0 @@
-package com.zetung.gifsgiver.model
-
-data class Gif(
-    val url: String
-)
